hpc/mpi_master: Adjust serial prompt
This needs to be done to get user serial terminal working again,
likely due: 2fcbb2b9f ("select_serial_terminal(): Update prompt")

Guarded by $OpenQA::Isotovideo::Interface::version ge 35 (changes in
20d0119b ("backend/qemu.pm: Add virtio_console_user fifo name")).

+++ tests/hpc/mpi_master.pm
@@ -10,13 +10,15 @@
 
 use Mojo::Base qw(hpcbase hpc::utils), -signatures;
 use testapi;
-use serial_terminal 'select_serial_terminal';
+use serial_terminal qw(select_serial_terminal select_user_serial_terminal);
 use lockapi;
 use utils;
 use registration;
 use version_utils 'is_sle';
 use Utils::Logging 'export_logs';
 use hpc::formatter;
+use isotovideo;
+
 use POSIX 'strftime';
 
 sub run ($self) {
@@ -30,14 +32,17 @@ sub run ($self) {
         bin => '/home/bernhard/bin',
         hpc_lib => '/usr/lib/hpc',
     );
+    my $user_virtio_fixed = isotovideo::get_version() >= 35;
+    my $prompt = $user_virtio_fixed ? $testapi::username . '@' . get_required_var('HOSTNAME') . ':~> ' : undef;
+
     script_run("sudo -u $testapi::username mkdir -p $exports_path{bin}");
     zypper_call("in $mpi-gnu-hpc $mpi-gnu-hpc-devel python3-devel");
     my $need_restart = $self->setup_scientific_module();
     $self->relogin_root if $need_restart;
     $self->setup_nfs_server(\%exports_path);
 
-    type_string('pkill -u root', lf => 1);
-    select_serial_terminal(0);
+    type_string('pkill -u root', lf => 1) unless $user_virtio_fixed;
+    select_user_serial_terminal($prompt);
     # for <15-SP2 the openmpi2 module is named simply openmpi
     $mpi = 'openmpi' if ($mpi =~ /openmpi2|openmpi3|openmpi4/);
 
@@ -59,8 +64,8 @@ sub run ($self) {
     systemctl 'restart nfs-server';
     # And login as normal user to run the tests
     # NOTE: This behaves weird. Need another solution apparently
-    type_string('pkill -u root');
-    select_serial_terminal(0);
+    type_string('pkill -u root') unless $user_virtio_fixed;
+    select_user_serial_terminal($prompt);
     # load mpi after all the relogins
     assert_script_run "module load gnu $mpi";
     script_run "module av";
